backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.7

# AI / ML
openai==1.35.0
tiktoken==0.7.0
sentence-transformers==3.0.1
transformers==4.42.3
torch==2.3.1

# Vector Store
chromadb==0.5.3

# Document Parsing
pypdf==4.2.0
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==0.6.23
beautifulsoup4==4.12.3
markdownify==0.13.1
chardet==5.2.0

# Search
rank-bm25==0.2.2
jieba==0.42.1

# Utils
pydantic==2.8.0
pydantic-settings==2.3.4
loguru==0.7.2
aiofiles==24.1.0
httpx==0.27.0
numpy==1.26.4
python-dotenv==1.0.0
python-magic==0.4.27